refactor: extract address screening logic into a hook
