+++ packages/arb-token-bridge-ui/src/hooks/useNetworksAndSigners.tsx
@@ -31,14 +31,12 @@ import { getWagmiChain } from '../util/wagmi/getWagmiChain'
 import { useArbQueryParams } from './useArbQueryParams'
 import { trackEvent } from '../util/AnalyticsUtils'
 
-import { ApiResponseSuccess } from '../pages/api/screenings'
 import { TOS_LOCALSTORAGE_KEY } from '../constants'
 
 export enum UseNetworksAndSignersStatus {
   LOADING = 'loading',
   NOT_CONNECTED = 'not_connected',
   NOT_SUPPORTED = 'not_supported',
-  BLOCKED = 'blocked',
   CONNECTED = 'connected'
 }
 
@@ -48,9 +46,6 @@ export type UseNetworksAndSignersLoadingOrErrorStatus =
 
 export type UseNetworksAndSignersNotSupportedStatus =
   UseNetworksAndSignersStatus.NOT_SUPPORTED
-
-export type UseNetworksAndSignersBlockedStatus =
-  UseNetworksAndSignersStatus.BLOCKED
 
 const defaultStatus =
   typeof window.ethereum === 'undefined'
@@ -64,11 +59,6 @@ export type UseNetworksAndSignersLoadingOrErrorResult = {
 export type UseNetworksAndSignersNotSupportedResult = {
   status: UseNetworksAndSignersStatus.NOT_SUPPORTED
   chainId: number // the current unsupported chainId which is connected to UI
-}
-
-export type UseNetworksAndSignersBlockedResult = {
-  status: UseNetworksAndSignersStatus.BLOCKED
-  address: string
 }
 
 export type UseNetworksAndSignersConnectedResult = {
@@ -89,7 +79,6 @@ export type UseNetworksAndSignersResult =
   | UseNetworksAndSignersLoadingOrErrorResult
   | UseNetworksAndSignersConnectedResult
   | UseNetworksAndSignersNotSupportedResult
-  | UseNetworksAndSignersBlockedResult
 
 export const NetworksAndSignersContext = createContext<
   UseNetworksAndSignersConnectedResult | undefined
@@ -110,7 +99,6 @@ export function useNetworksAndSigners() {
 export type FallbackProps =
   | { status: UseNetworksAndSignersLoadingOrErrorStatus }
   | { status: UseNetworksAndSignersNotSupportedStatus; chainId: number }
-  | { status: UseNetworksAndSignersStatus.BLOCKED; address: string }
 
 export type NetworksAndSignersProviderProps = {
   /**
@@ -160,27 +148,6 @@ function getWalletName(connectorName: string): ProviderName {
   }
 }
 
-async function isBlocked(address: `0x${string}`): Promise<boolean> {
-  if (
-    process.env.NODE_ENV !== 'production' ||
-    process.env.NEXT_PUBLIC_IS_E2E_TEST
-  ) {
-    return false
-  }
-
-  const searchParams = new URLSearchParams({ address })
-  const response = await fetch('/api/screenings?' + searchParams, {
-    method: 'GET',
-    headers: { 'Content-Type': 'application/json' }
-  })
-
-  if (!response.ok) {
-    return false
-  }
-
-  return ((await response.json()) as ApiResponseSuccess).blocked
-}
-
 export function NetworksAndSignersProvider(
   props: NetworksAndSignersProviderProps
 ): JSX.Element {
@@ -218,16 +185,6 @@ export function NetworksAndSignersProvider(
   // TODO: Don't run all of this when an account switch happens. Just derive signers from networks?
   const update = useCallback(async () => {
     if (!address || !chain) {
-      return
-    }
-    const blocked = await isBlocked(address)
-
-    if (blocked) {
-      setResult({
-        status: UseNetworksAndSignersStatus.BLOCKED,
-        address: address as string
-      })
-      trackEvent('Address Block')
       return
     }
 
@@ -350,10 +307,6 @@ export function NetworksAndSignersProvider(
   }, [chain, update])
 
   if (result.status !== UseNetworksAndSignersStatus.CONNECTED) {
-    if (result.status === UseNetworksAndSignersStatus.BLOCKED) {
-      return props.fallback(result)
-    }
-
     const fallbackProps =
       result.status === UseNetworksAndSignersStatus.NOT_SUPPORTED
         ? {
